requirements.txt
imageio
Pillow
python-magic